Created "wessex-main"
This is quite a big one.  I've created "wessex-main" in the repository,
and uploaded it onto "http://argles.org/wessex-scd/wessex-main/".  It's
a first go at updating the wessex-scd page idea to HTML5 and CSS3.
There's quite a way to go, but it should be as compatible as possible
with as many variants as possible.  And hopefully should be
future-proofed.