More compatibility updates
Rather a lot in one commit!  The footer is now a bit tidier and the
"Return to Top" button is sorted out.  Formatting for narrow screen
widths has been tidied up.  The W3 compliance icons have been sorted.
The map page has been sorted - mps are now encased in a
<figure></figure> to get the formatting right.