Created new Winchester club site
Created a new version of the Winchester club site.  The used this to:
get table formatting right in the new version; sort out the "thistle"
graphic; and sort out the <h2> heading (for some reason, Winchester uses
<h2>, not <h3>.  But this is different from the <h2> used as a site
sub-heading...
Also found one or two errors/omissions in template.php which are now
correected.